Split scopes into scope object and data object
Previously a scope object was a regular object with two extra properties tacked on: strict and parentScope.  Issue #183 proved that sometimes one needs multiple scopes with different parentScope values.  This PR creates a scope object with the above two properties, as well as a third property (‘.object’) that points to the regular object.  This allows two scopes to have different parentScopes, while sharing the same object for variables.

Minor API change for advanced threading and serialization.

Resolves #183