Async callback with no args should return
Existing usage includes calling the callback with no arguments, intending it to return undefined.  Preserve this behaviour by treating zero and one arguments equivalently, and only throwing if it is called with two or more.

(This will break any existing usage which calls the callback with more than one argument, but that ought to be fairly uncommon!)