Double speed of serialization.
Change objectHunt_ to iterative rather than recursive.  This makes no measurable performance difference, but does set up for the next change.

Add a Set to verify the uniqueness of objects, rather than using array's indexOf.  This doubles the performance of serialization, but only works in ES6+.  ES5 falls back to the indexOf approach.