Catchable readonly variables
Readonly object properties throw when they are set in strict mode, and that throw is catchable by the user.  However readonly variables would throw uncatchable native errors when set.

Until now, there were no readonly variables (stuff like Infinity are properties on the global object).  But with the introduction of function names as variables, that changes.